Test where the roi flag is thread specific.
Advantages
We can set the roi inside a parallel region to trace only the contents
of the thread (not the omp admin around it):

    #pragma omp parallel
    {
        atomic_trace::start_roi();
        sum += 1;
        atomic_trace::end_roi();
    }

Disadvantage
Setting the roi at the beginning of main will trace all threads with the
old global roi like this:

    atomic_trace::start_roi();

    #pragma omp parallel
    {
        sum += 1;
    }

    atomic_trace::end_roi();

With this per thread roi we only get the trace of the main thread. Which
is normally not what you want. To get all thread tracing we need to
enable the roi for every thread as shown in the first example above.